% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupsmigration_functions.R
\name{archive.insert}
\alias{archive.insert}
\title{Inserts a new mail into the archive of the Google group.}
\usage{
archive.insert(groupId)
}
\arguments{
\item{groupId}{The group ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.groups.migration
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.groups.migration)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/groups-migration/}{Google Documentation}
}
