% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupsmigration_objects.R
\name{Groups}
\alias{Groups}
\title{Groups Migration API Objects 
Groups Migration Api.}
\usage{
Groups(responseCode = NULL)
}
\arguments{
\item{responseCode}{The status of the insert request}
}
\value{
Groups object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:15:13
filename: /Users/mark/dev/R/autoGoogleAPI/googlegroupsmigrationv1.auto/R/groupsmigration_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Groups Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON response template for groups migration API.
}

