% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupsmigration_functions.R
\docType{package}
\name{groupsmigration_googleAuthR}
\alias{groupsmigration_googleAuthR}
\alias{groupsmigration_googleAuthR-package}
\title{Groups Migration API
Groups Migration Api.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:15:13
filename: /Users/mark/dev/R/autoGoogleAPI/googlegroupsmigrationv1.auto/R/groupsmigration_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/apps.groups.migration
}
}

